+++ app/models/__init__.py
@@ -1,8 +1,6 @@
-from .user_model import ClientModel, EmployeeModel
 from .offer_model import OfferLinkModel
 from .order_model import OrderModel, OrderDetailModel
 from .service_model import ServiceNameModel
-# from category.model_category import CategoryModel
 
 # User [Client, Master] (email)
 # -------------
