+++ app/services/client_service.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from models import ClientModel
+from entities.users.model_user import ClientModel
 from services.base_service import AbstractService
