+++ app/tests/test_employee.py
@@ -1,6 +1,6 @@
 from typing import Type
 
-from models import EmployeeModel
+from entities.users.model_user import EmployeeModel
 from tests.abstract_user import UserAbstract
 from schemas.user_schemas import EmployeeInSchema, EmployeeInOptionalSchema
 from services.stub_init_service import LAST_NAMES
